Add module for helper methods
